+++ demo/loader.js
@@ -10,8 +10,10 @@ const isUsingTopsortBlocks = true;
 const isUsingCustomProps = false;
 const useTopsortedProducts = true;
 const customPromoteTargetClassName = "my-custom-promote-target";
-const CENTRAL_SERVICES_URL = "https://api.topsort.com/";
-const AUTH_BASE_URL = "https://buildingblocks.api.topsort.ai";
+const CENTRAL_SERVICES_URL = "http://localhost:3003";
+// const CENTRAL_SERVICES_URL = "https://api.topsort.com/";
+const AUTH_BASE_URL = "http://localhost:8000";
+// const AUTH_BASE_URL = "https://buildingblocks.api.topsort.ai";
 
 function getNewElement(selector) {
   const proto = document.querySelector(selector);
