+++ src/components/CampaignDetails/CampaignEnded.tsx
@@ -3,7 +3,7 @@ import { Button } from "@components/Button";
 import { Icon } from "@components/Icon";
 import { ModalContent, ModalHeading } from "@components/Modal";
 import { CampaignSummary } from "@components/common";
-import { useProductPromotion } from "@context";
+import { usePromotionContext } from "@context";
 import { h, FunctionComponent, Fragment } from "preact";
 
 import { Metrics } from "./Metrics";
@@ -11,7 +11,7 @@ import { Metrics } from "./Metrics";
 export const CampaignEnded: FunctionComponent<{ campaign: Campaign }> = ({
   campaign,
 }) => {
-  const { dispatch } = useProductPromotion();
+  const { dispatch } = usePromotionContext();
   const title = (
     <Fragment>
       <Icon size={32} name="tick-circle" className="ts-text-success" />
