+++ src/components/CampaignDetails/Details.tsx
@@ -1,7 +1,7 @@
 import { Campaign } from "@api/types";
 import { CampaignBudget, CampaignSummary } from "@components/common";
 import { MS_PER_DAY } from "@constants";
-import { useProductPromotion } from "@context";
+import { usePromotionContext } from "@context";
 import { h, FunctionalComponent } from "preact";
 import { useMemo } from "preact/hooks";
 
@@ -10,7 +10,7 @@ import { Metrics } from "./Metrics";
 export const Details: FunctionalComponent<{
   campaign: Campaign;
 }> = ({ campaign }) => {
-  const { dispatch } = useProductPromotion();
+  const { dispatch } = usePromotionContext();
   const days = useMemo(() => {
     const end = new Date(campaign.endDate);
